minor (and old) bugs fixed. minor changes in search paths for help and translation files.

+++ pcbnew/set_grid.cpp
@@ -34,10 +34,10 @@ void WinEDA_PcbGridFrame::AcceptPcbOptions(wxCommandEvent& event)
 	m_Parent->GetScreen()->AddGrid( g_UserGrid, g_UserGrid_Unit, ID_POPUP_GRID_USER );
 
 	EndModal(1);
-    
+
     // If the user grid is the current selection , ensure grid size value = new user grid value
     int ii = m_Parent->m_SelGridBox->GetSelection();
-    if ( ii == m_Parent->m_SelGridBox->GetCount() - 1 )
+    if ( ii == (int)(m_Parent->m_SelGridBox->GetCount() - 1) )
         m_Parent->GetScreen()->SetGrid( ID_POPUP_GRID_USER );
 
 	m_Parent->DrawPanel->Refresh( TRUE );
